Use common decorator to inhibit word completions
