+++ plugins/lib/__init__.py
@@ -1,3 +1,4 @@
+import functools
 import logging
 
 import sublime
@@ -19,3 +20,13 @@ def package_settings():
 
 def get_setting(key, default=None):
     return package_settings().get(key, default)
+
+
+def inhibit_word_completions(func):
+    """Decorator that inhibits ST's word completions if non-None value is returned."""
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        ret = func(*args, **kwargs)
+        return (ret, sublime.INHIBIT_WORD_COMPLETIONS) if ret is not None else None
+
+    return wrapper
